requirements.txt
fastapi
uvicorn
python-multipart
passlib[argon2]
python-jose
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-community
langchain-core
langchain-text-splitters
pdfplumber
sentence-transformers
faiss-cpu
numpy
openai
psycopg2-binary
sqlalchemy
streamlit
requests